+++ src/libAtomVM/jit.h
@@ -186,6 +186,7 @@ enum TrapAndLoadResult
 #define JIT_ARCH_ARMV6M 3
 #define JIT_ARCH_RISCV32 4
 #define JIT_ARCH_RISCV64 5
+#define JIT_ARCH_ARM32 6
 
 #define JIT_VARIANT_PIC 1
 #define JIT_VARIANT_FLOAT32 2
@@ -202,7 +203,10 @@ enum TrapAndLoadResult
 #define JIT_JUMPTABLE_ENTRY_SIZE 4
 #endif
 
-#ifdef __arm__
+#if defined(__arm__) && defined(AVM_JIT_ARM32)
+#define JIT_ARCH_TARGET JIT_ARCH_ARM32
+#define JIT_JUMPTABLE_ENTRY_SIZE 8
+#elif defined(__arm__)
 #define JIT_ARCH_TARGET JIT_ARCH_ARMV6M
 #define JIT_JUMPTABLE_ENTRY_SIZE 12
 #endif