+++ src/libAtomVM/nifs.c
@@ -6250,6 +6250,8 @@ static term nif_jit_backend_module(Context *ctx, int argc, term argv[])
     return JIT_RISCV32_ATOM;
 #elif JIT_ARCH_TARGET == JIT_ARCH_RISCV64
     return JIT_RISCV64_ATOM;
+#elif JIT_ARCH_TARGET == JIT_ARCH_ARM32
+    return JIT_ARM32_ATOM;
 #else
 #error Unknown JIT target
 #endif